+++ client/views/posts/post_submit.js
@@ -1,54 +1,42 @@
 AutoForm.hooks({
   submitPostForm: {
-    onSubmit: function(insertDoc, updateDoc, currentDoc) {
 
-      var properties = insertDoc;
-      var submit = this;
+    before: {
+      submitPost: function(doc, template) {
 
-      // ------------------------------ Checks ------------------------------ //
+        template.$('button[type=submit]').addClass('loading');
 
-      if (!Meteor.user()) {
-        flashMessage(i18n.t('you_must_be_logged_in'), 'error');
-        return false;
-      }
+        var post = doc;
 
-      // ------------------------------ Callbacks ------------------------------ //
+        // ------------------------------ Checks ------------------------------ //
 
-      // run all post submit client callbacks on properties object successively
-      properties = postSubmitClientCallbacks.reduce(function(result, currentFunction) {
-          return currentFunction(result);
-      }, properties);
-
-      // console.log(properties)
-
-      // ------------------------------ Insert ------------------------------ //
-      Meteor.call('submitPost', properties, function(error, post) {
-        if(error){
-          submit.done(error);
-        }else{
-          // note: find a way to do this in onSuccess instead?
-          trackEvent("new post", {'postId': post._id});
-          if (post.status === STATUS_PENDING) {
-            flashMessage(i18n.t('thanks_your_post_is_awaiting_approval'), 'success');
-          }
-          Router.go('post_page', {_id: post._id});
-          submit.done();
+        if (!Meteor.user()) {
+          flashMessage(i18n.t('you_must_be_logged_in'), 'error');
+          return false;
         }
-      });
 
-      return false
+        // ------------------------------ Callbacks ------------------------------ //
+
+        // run all post submit client callbacks on properties object successively
+        post = postSubmitClientCallbacks.reduce(function(result, currentFunction) {
+            return currentFunction(result);
+        }, post);
+
+        return post;
+      }
     },
 
-    onSuccess: function(operation, result, template) {
-      // not used right now because I can't find a way to pass the "post" object to this callback
-      // console.log(post)
-      // trackEvent("new post", {'postId': post._id});
-      // if(post.status === STATUS_PENDING)
-      //   throwError('Thanks, your post is awaiting approval.');
-      // Router.go('/posts/'+post._id);
+    onSuccess: function(operation, post, template) {      
+      template.$('button[type=submit]').removeClass('loading');
+      trackEvent("new post", {'postId': post._id});
+      if (post.status === STATUS_PENDING) {
+        flashMessage(i18n.t('thanks_your_post_is_awaiting_approval'), 'success');
+      }
+      Router.go('post_page', {_id: post._id});
     },
 
     onError: function(operation, error, template) {
+      template.$('button[type=submit]').removeClass('loading');
       flashMessage(error.message.split('|')[0], 'error'); // workaround because error.details returns undefined
       clearSeenMessages();
       // $(e.target).removeClass('disabled');
@@ -58,12 +46,5 @@ AutoForm.hooks({
       }
     }
 
-    // Called at the beginning and end of submission, respectively.
-    // This is the place to disable/enable buttons or the form,
-    // show/hide a "Please wait" message, etc. If these hooks are
-    // not defined, then by default the submit button is disabled
-    // during submission.
-    // beginSubmit: function(formId, template) {},
-    // endSubmit: function(formId, template) {}
   }
 });