+++ collections/posts.js
@@ -251,7 +251,8 @@ getPostProperties = function (post) {
 
 // default status for new posts
 getDefaultPostStatus = function (user) {
-  if (isAdmin(user) || !getSetting('requirePostsApproval', false)) {
+  var hasAdminRights = typeof user === 'undefined' ? false : isAdmin(user);
+  if (hasAdminRights || !getSetting('requirePostsApproval', false)) {
     // if user is admin, or else post approval is not required
     return STATUS_APPROVED
   } else {
@@ -273,17 +274,24 @@ getPostLink = function (post) {
   return !!post.url ? getOutgoingUrl(post.url) : getPostPageUrl(post);
 };
 
-checkForPostsWithSameUrl = function (url) {
+// we need the current user so we know who to upvote the existing post as
+checkForPostsWithSameUrl = function (url, currentUser) {
 
   // check that there are no previous posts with the same link in the past 6 months
   var sixMonthsAgo = moment().subtract(6, 'months').toDate();
   var postWithSameLink = Posts.findOne({url: url, postedAt: {$gte: sixMonthsAgo}});
 
   if(typeof postWithSameLink !== 'undefined'){
-    Meteor.call('upvotePost', postWithSameLink);
+    upvoteItem(Posts, postWithSameLink, currentUser);
+
     // note: error.details returns undefined on the client, so add post ID to reason
     throw new Meteor.Error('603', i18n.t('this_link_has_already_been_posted') + '|' + postWithSameLink._id, postWithSameLink._id);
   }
+}
+
+// when on a post page, return the current post
+currentPost = function () {
+  return Posts.findOne(Router.current().data()._id);
 }
 
 // ------------------------------------------------------------------------------------------- //
@@ -303,6 +311,86 @@ Posts.before.update(function (userId, doc, fieldNames, modifier, options) {
   }
 });
 
+postAfterSubmitMethodCallbacks.push(function (post) {
+
+  var userId = post.userId,
+      postAuthor = Meteor.users.findOne(userId);
+
+  // increment posts count
+  Meteor.users.update({_id: userId}, {$inc: {postCount: 1}});
+  upvoteItem(Posts, post, postAuthor);
+  
+  return post;
+
+});
+
+// ------------------------------------------------------------------------------------------- //
+// --------------------------------------- Submit Post --------------------------------------- //
+// ------------------------------------------------------------------------------------------- //
+
+submitPost = function (post) {
+
+  var userId = post.userId, // at this stage, a userId is expected
+      user = Meteor.users.findOne(userId);
+
+  // ------------------------------ Checks ------------------------------ //
+
+  // check that a title was provided
+  if(!post.title)
+    throw new Meteor.Error(602, i18n.t('please_fill_in_a_title'));
+
+  // check that there are no posts with the same URL
+  if(!!post.url)
+    checkForPostsWithSameUrl(post.url, user);
+
+  // ------------------------------ Properties ------------------------------ //
+
+  defaultProperties = {
+    createdAt: new Date(),
+    author: getDisplayNameById(userId),
+    upvotes: 0,
+    downvotes: 0,
+    commentCount: 0,
+    clickCount: 0,
+    viewCount: 0,
+    baseScore: 0,
+    score: 0,
+    inactive: false,
+    sticky: false,
+    status: getDefaultPostStatus(),
+    postedAt: new Date()
+  };
+
+  post = _.extend(defaultProperties, post);
+
+  // clean up post title
+  post.title = cleanUp(post.title);
+
+  // ------------------------------ Callbacks ------------------------------ //
+
+  // run all post submit server callbacks on post object successively
+  post = postSubmitMethodCallbacks.reduce(function(result, currentFunction) {
+      return currentFunction(result);
+  }, post);
+
+  // -------------------------------- Insert ------------------------------- //
+
+  post._id = Posts.insert(post);
+
+  // --------------------- Server-Side Async Callbacks --------------------- //
+
+  if (Meteor.isServer) {
+    Meteor.defer(function () { // use defer to avoid holding up client
+      // run all post submit server callbacks on post object successively
+      post = postAfterSubmitMethodCallbacks.reduce(function(result, currentFunction) {
+          return currentFunction(result);
+      }, post);
+    });
+  }
+
+  return post;
+}
+
 // ------------------------------------------------------------------------------------------- //
 // ----------------------------------------- Methods ----------------------------------------- //
 // ------------------------------------------------------------------------------------------- //
@@ -313,15 +401,19 @@ postViews = [];
 Meteor.methods({
 
   submitPost: function(post){
-    var title = cleanUp(post.title),
-        body = post.body,
-        userId = this.userId,
-        user = Meteor.users.findOne(userId),
-        timeSinceLastPost=timeSinceLast(user, Posts),
-        numberOfPostsInPast24Hours=numberOfItemsInPast24Hours(user, Posts),
-        postInterval = Math.abs(parseInt(getSetting('postInterval', 30))),
-        maxPostsPer24Hours = Math.abs(parseInt(getSetting('maxPostsPerDay', 30))),
-        postId = '';
+
+    // required properties:
+    // title
+
+    // optional properties
+    // URL
+    // body
+    // categories
+    // thumbnailUrl
+
+    // NOTE: the current user and the post author user might be two different users!
+    var user = Meteor.user(),
+        hasAdminRights = isAdmin(user);
 
     // ------------------------------ Checks ------------------------------ //
 
@@ -329,108 +421,57 @@ Meteor.methods({
     if (!user || !canPost(user))
       throw new Meteor.Error(601, i18n.t('you_need_to_login_or_be_invited_to_post_new_stories'));
 
-    // check that user provided a title
-    if(!post.title)
-      throw new Meteor.Error(602, i18n.t('please_fill_in_a_title'));
-
-    // check that there are no posts with the same URL
-    if(!!post.url)
-      checkForPostsWithSameUrl(post.url);
-
     // --------------------------- Rate Limiting -------------------------- //
 
-    if(!isAdmin(Meteor.user())){
+    if(!hasAdminRights){
+    
+      var timeSinceLastPost=timeSinceLast(user, Posts),
+        numberOfPostsInPast24Hours=numberOfItemsInPast24Hours(user, Posts),
+        postInterval = Math.abs(parseInt(getSetting('postInterval', 30))),
+        maxPostsPer24Hours = Math.abs(parseInt(getSetting('maxPostsPerDay', 30)));
+
       // check that user waits more than X seconds between posts
-      if(!this.isSimulation && timeSinceLastPost < postInterval)
+      if(timeSinceLastPost < postInterval)
         throw new Meteor.Error(604, i18n.t('please_wait')+(postInterval-timeSinceLastPost)+i18n.t('seconds_before_posting_again'));
 
       // check that the user doesn't post more than Y posts per day
-      if(!this.isSimulation && numberOfPostsInPast24Hours > maxPostsPer24Hours)
+      if(numberOfPostsInPast24Hours > maxPostsPer24Hours)
         throw new Meteor.Error(605, i18n.t('sorry_you_cannot_submit_more_than')+maxPostsPer24Hours+i18n.t('posts_per_day'));
+    
     }
 
     // ------------------------------ Properties ------------------------------ //
 
-    // Basic Properties
-    properties = {
-      title: title,
-      body: body,
-      userId: userId,
-      author: getDisplayNameById(userId),
-      upvotes: 0,
-      downvotes: 0,
-      commentCount: 0,
-      clickCount: 0,
-      viewCount: 0,
-      baseScore: 0,
-      score: 0,
-      inactive: false
-    };
+    // admin-only properties
+    // status
+    // postedAt
+    // userId
+    // sticky (default to false)
 
-    // UserId
-    if(isAdmin(Meteor.user()) && !!post.userId){ // only let admins post as other users
-      properties.userId = post.userId;
+    // if user is not admin, clear restricted properties
+    if (!hasAdminRights) {
+      delete post.status;
+      delete post.postedAt;
+      delete post.userId;
+      delete post.sticky;
     }
 
-    // Status
-    if(!!post.status && isAdmin(Meteor.user())){
-      // if a custom status has been set, and user is admin, use that
-      properties.status = post.status;
-    }else{
-      // else use default status
-      properties.status = getDefaultPostStatus(Meteor.user());
+    // if no post status has been set, set it now
+    if (!post.status) {
+      post.status = getDefaultPostStatus(user);
     }
 
-    // CreatedAt
-    properties.createdAt = new Date();
-
-    // PostedAt
-    if(properties.status == 2){ // only set postedAt if post is approved
-      if(isAdmin(Meteor.user()) && !!post.postedAt){ // if user is admin and a custom postDate has been set
-        properties.postedAt = post.postedAt;
-      }else{ // else use current time
-        properties.postedAt = new Date();
-      }
+    // if no userId has been set, default to current user id
+    if (!post.userId) {
+      post.userId = user._id
     }
-
-    post = _.extend(post, properties);
-
-    // ------------------------------ Callbacks ------------------------------ //
-
-    // run all post submit server callbacks on post object successively
-    post = postSubmitMethodCallbacks.reduce(function(result, currentFunction) {
-        return currentFunction(result);
-    }, post);
-
-    // ------------------------------ Insert ------------------------------ //
-
-    // console.log(post)
-    post._id = Posts.insert(post);
-
-    // ------------------------------ Callbacks ------------------------------ //
-
-    // run all post submit server callbacks on post object successively
-    post = postAfterSubmitMethodCallbacks.reduce(function(result, currentFunction) {
-        return currentFunction(result);
-    }, post);
-
-    // ------------------------------ After Insert ------------------------------ //
-
-    // increment posts count
-    Meteor.users.update({_id: userId}, {$inc: {postCount: 1}});
-
-    var postAuthor =  Meteor.users.findOne(post.userId);
-
-    Meteor.call('upvotePost', post, postAuthor);
-
-    return post;
+   
+    return submitPost(post);
   },
 
-  editPost: function (postId, updateObject) {
+  editPost: function (post, modifier, postId) {
 
     var user = Meteor.user();
-
-    // console.log(updateObject)
 
     // ------------------------------ Checks ------------------------------ //
 
@@ -440,23 +481,21 @@ Meteor.methods({
 
     // ------------------------------ Callbacks ------------------------------ //
 
-    // run all post submit server callbacks on updateObject successively
-    updateObject = postEditMethodCallbacks.reduce(function(result, currentFunction) {
+    // run all post submit server callbacks on modifier successively
+    modifier = postEditMethodCallbacks.reduce(function(result, currentFunction) {
         return currentFunction(result);
-    }, updateObject);
-
-    console.log(updateObject)
+    }, modifier);
 
     // ------------------------------ Update ------------------------------ //
 
-    Posts.update(postId, updateObject);
+    Posts.update(postId, modifier);
 
     // ------------------------------ Callbacks ------------------------------ //
 
-    // run all post submit server callbacks on updateObject successively
-    updateObject = postAfterEditMethodCallbacks.reduce(function(result, currentFunction) {
+    // run all post submit server callbacks on modifier object successively
+    modifier = postAfterEditMethodCallbacks.reduce(function(result, currentFunction) {
         return currentFunction(result);
-    }, updateObject);
+    }, modifier);
 
     // ------------------------------ After Update ------------------------------ //