+++ packages/telescope-datetimepicker/package.js
@@ -7,7 +7,7 @@ Package.describe({
 Package.onUse(function(api) {
   api.use('templating@1.0.0');
   api.use('blaze@2.0.0');
-  api.use('aldeed:autoform@4.0.0');
+  api.use('aldeed:autoform');
   api.use('fourseven:scss');
   // api.use('jquery');
   // api.use('tsega:bootstrap3-datetimepicker');