+++ packages/telescope-post-by-feed/lib/client/routes.js
@@ -1,0 +1,21 @@
+adminNav.push({
+  route: 'feeds',
+  label: 'Feeds',
+  description: 'import_new_posts_from_feeds'
+});
+
+Meteor.startup(function () {
+
+  Router.onBeforeAction(Router._filters.isAdmin, {only: ['feeds']});
+  
+  // RSS Urls Admin
+
+  Router.route('/feeds', {
+    name: 'feeds',
+    waitOn: function() {
+      return Meteor.subscribe('feeds');
+    },
+    template: getTemplate('feeds')
+  });
+
+});