+++ packages/telescope-post-by-feed/lib/server/cron.js
@@ -1,0 +1,9 @@
+SyncedCron.add({
+  name: 'Post by RSS feed',
+  schedule: function(parser) {
+    return parser.text('every 30 minutes');
+  }, 
+  job: function() {
+    fetchFeeds();
+  }
+});