+++ packages/telescope-post-by-feed/lib/server/publications.js
@@ -1,0 +1,6 @@
+Meteor.publish('feeds', function() {
+  if(isAdminById(this.userId)){
+    return Feeds.find();
+  }
+  return [];
+});